chore: first part of structure change
