+++ run.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-from watcher.web import main
+from watcher.app import main
 
 if __name__ == "__main__":
     main()
