+++ tests/test_web.py
@@ -4,37 +4,37 @@ import responses
 from fastapi.testclient import TestClient
 from test_in_memory_state import task_template
 
-from watcher.web import app
+from watcher.app import app
 
 client = TestClient(app)
 api_path = "/api/v1/tasks"
 
 
 def responses_configuration():
+    responses.add(method=responses.GET, url="https://argocd.example.com", status=200)
     responses.add(
         method=responses.GET,
-        url='https://argocd.example.com',
-        status=200
+        url="https://argocd.example.com/api/v1/applications/test_app",
+        json={
+            "status": {
+                "summary": {"images": ["example:latest"]},
+                "sync": {"status": "Synced"},
+                "health": {"status": "Healthy"},
+            }
+        },
+        status=200,
     )
     responses.add(
         method=responses.GET,
-        url='https://argocd.example.com/api/v1/applications/test_app',
-        json={'status': {
-            'summary': {'images': ['example:latest']},
-            'sync': {'status': 'Synced'},
-            'health': {'status': 'Healthy'}
-        }},
-        status=200
-    )
-    responses.add(
-        method=responses.GET,
-        url='https://argocd.example.com/api/v1/applications/example',
-        json={'status': {
-            'summary': {'images': ['example:latest']},
-            'sync': {'status': 'Synced'},
-            'health': {'status': 'Healthy'}
-        }},
-        status=200
+        url="https://argocd.example.com/api/v1/applications/example",
+        json={
+            "status": {
+                "summary": {"images": ["example:latest"]},
+                "sync": {"status": "Synced"},
+                "health": {"status": "Healthy"},
+            }
+        },
+        status=200,
     )
 
 
@@ -44,27 +44,27 @@ def test_healthz():
 
     responses.add(
         method=responses.GET,
-        url='https://argocd.example.com/api/v1/session/userinfo',
+        url="https://argocd.example.com/api/v1/session/userinfo",
         json={"loggedIn": True},
-        status=200
+        status=200,
     )
 
     response = client.get("/healthz")
 
     assert response.status_code == 200
-    assert response.json()['status'] == "up"
+    assert response.json()["status"] == "up"
 
     responses.add(
         method=responses.GET,
-        url='https://argocd.example.com/api/v1/session/userinfo',
+        url="https://argocd.example.com/api/v1/session/userinfo",
         json={},
-        status=503
+        status=503,
     )
 
     response = client.get("/healthz")
 
     assert response.status_code == 503
-    assert response.json()['status'] == "down"
+    assert response.json()["status"] == "down"
 
 
 @responses.activate
@@ -73,39 +73,41 @@ def test_add_task():
     response = client.post(api_path, json=task_template)
 
     assert response.status_code == 202
-    assert response.json()['status'] == "accepted"
-    assert len(response.json()['id']) == 36
+    assert response.json()["status"] == "accepted"
+    assert len(response.json()["id"]) == 36
 
 
 @responses.activate
 def test_get_task_status():
     responses_configuration()
-    task_id = client.post(api_path, json=task_template).json()['id']
+    task_id = client.post(api_path, json=task_template).json()["id"]
     response = client.get(f"{api_path}/{task_id}")
 
     assert response.status_code == 200
-    assert response.json()['status'] == "deployed"
+    assert response.json()["status"] == "deployed"
 
 
 @responses.activate
 def test_get_state_with_filter():
     responses_configuration()
     target_task = task_template.copy()
-    target_task['app'] = 'example'
+    target_task["app"] = "example"
 
     client.post(api_path, json=task_template)
     client.post(api_path, json=task_template)
     client.post(api_path, json=target_task)
 
-    response = client.get(api_path, params={"from_timestamp": time() - 60, "app": "example"})
+    response = client.get(
+        api_path, params={"from_timestamp": time() - 60, "app": "example"}
+    )
 
     assert response.status_code == 200
     assert len(response.json()) == 1
-    assert response.json()[0]['app'] == "example"
+    assert response.json()[0]["app"] == "example"
 
 
 def test_get_app_list():
     response = client.get("/api/v1/apps")
 
     assert len(response.json()) == 2
-    assert set(response.json()) == {'test_app', 'example'}
+    assert set(response.json()) == {"test_app", "example"}
